+++ tesseract/cppbuild.sh
@@ -8,16 +8,19 @@ if [[ -z "$PLATFORM" ]]; then
 fi
 
 if [[ $PLATFORM == windows* ]]; then
-    TESSERACT_VERSION=3.02.02
+    TESSERACT_VERSION=3.03-0.1.rc1
     [[ $PLATFORM == *64 ]] && BITS=64 || BITS=32
-    download https://tesseract-ocr.googlecode.com/files/tesseract-$TESSERACT_VERSION-win$BITS-lib-include-dirs.zip tesseract-$TESSERACT_VERSION-win$BITS-lib-include-dirs.zip
+    download http://mirrors.kernel.org/fedora/releases/21/Everything/x86_64/os/Packages/m/mingw$BITS-tesseract-$TESSERACT_VERSION.fc21.noarch.rpm mingw$BITS-tesseract-$TESSERACT_VERSION.fc21.noarch.rpm
+
+    function extract {
+        /C/Program\ Files/7-Zip/7z x -y $1
+    }
+    extract mingw$BITS-tesseract-$TESSERACT_VERSION.fc21.noarch.rpm
 
     mkdir -p $PLATFORM
     cd $PLATFORM
-    unzip -o ../tesseract-$TESSERACT_VERSION-win$BITS-lib-include-dirs.zip
-    cd tesseract-$TESSERACT_VERSION-win$BITS-lib-include-dirs
-    mv include ..
-    mv lib ..
+    rm -Rf include lib bin
+    extract ../mingw$BITS-tesseract-$TESSERACT_VERSION.fc21.noarch.cpio
 else
     TESSERACT_VERSION=3.03
     download "https://drive.google.com/uc?export=download&id=0B7l10Bj_LprhSGN2bTYwemVRREU" tesseract-$TESSERACT_VERSION.tar.gz
@@ -60,8 +63,10 @@ case $PLATFORM in
         make install-strip
         ;;
     windows-x86)
+        mv usr/i686-w64-mingw32/sys-root/mingw/* .
         ;;
     windows-x86_64)
+        mv usr/x86_64-w64-mingw32/sys-root/mingw/* .
         ;;
     *)
         echo "Error: Platform \"$PLATFORM\" is not supported"